planners: astar: move the header in include/planners/astar
